Add logics to create a order

+++ source/BlossomAvenue.Infrastructure/Repositories/Orders/OrderRepository.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BlossomAvenue.Core.Carts;
 using BlossomAvenue.Core.Orders;
 using BlossomAvenue.Infrastructure.Database;
 using BlossomAvenue.Service.CustomExceptions;
@@ -18,72 +19,33 @@ namespace BlossomAvenue.Infrastructure.Repositories.Orders
         {
             _context = context;
         }
-        public async Task<bool> CreateOrder(Guid cartId, Guid userId)
+        public async Task<Order> CreateOrder(Cart cart, Order order)
         {
-            var cart = await _context.Carts
-                                     .Include(c => c.CartItems)
-                                     .FirstOrDefaultAsync(c => c.CartId == cartId);
 
-            if (cart == null || !cart.CartItems.Any())
+            var newOrder = (await _context.Orders.AddAsync(order)).Entity;
+            if (newOrder != null)
             {
-                throw new InvalidOperationException("Cart is empty or does not exist.");
-            }
-
-            var order = new Order
-            {
-                OrderId = Guid.NewGuid(),
-                UserId = userId,
-                AddressId = null,
-                CreatedAt = DateTime.UtcNow,
-                OrderStatus = "pending"
-            };
-
-            decimal? totalAmount = 0;
-
-            foreach (var cartItem in cart.CartItems)
-            {
-                var variation = await _context.Variations
-                                      .FirstOrDefaultAsync(v => v.VariationId == cartItem.VariationId);
-
-                if (variation == null)
+                _context.CartItems.RemoveRange(cart.CartItems);
+                foreach (var cartItem in cart.CartItems)
                 {
-                    throw new InvalidOperationException("Invalid variation for a cart item.");
+                    var variation = await _context.Variations.FirstOrDefaultAsync(v => v.VariationId == cartItem.VariationId);
+                    if (variation != null)
+                    {
+                        variation.Inventory -= cartItem.Quantity;
+                    }
                 }
-
-                // Calculate the total price
-                decimal itemPrice = variation.Price; // Assuming Variation has a Price property
-                decimal totalPrice = itemPrice * cartItem.Quantity;
-
-                totalAmount += totalPrice;
-
-                var orderItem = new OrderItem
-                {
-                    OrderItemsId = Guid.NewGuid(),
-                    OrderId = order.OrderId,
-                    ProductId = cartItem.ProductId,
-                    VariationId = cartItem.VariationId,
-                    Quantity = cartItem.Quantity,
-                    Price = totalPrice
-                };
-
-                order.OrderItems.Add(orderItem);
             }
-
-            order.TotalAmount = totalAmount;
-            _context.Orders.Add(order);
-
-            _context.CartItems.RemoveRange(cart.CartItems);
-
-            await _context.SaveChangesAsync();
-
-            return true;
+            if (await _context.SaveChangesAsync() > 0) return newOrder;
+            return null;
         }
 
         public async Task<Order> GetOrder(Guid orderId)
         {
             return await _context.Orders
-                                     .Include(o => o.OrderItems)
-                                     .FirstOrDefaultAsync(o => o.OrderId == orderId);
+                        .Include(o => o.OrderItems)
+                        .Include(o => o.AddressDetail)
+                        .ThenInclude(o => o.City)
+                        .FirstOrDefaultAsync(o => o.OrderId == orderId);
         }
 
         public async Task<bool> UpdateOrder(Guid orderId, string orderStatus)
@@ -96,7 +58,7 @@ namespace BlossomAvenue.Infrastructure.Repositories.Orders
                 throw new RecordNotFoundException("order");
             }
 
-            order.OrderStatus = orderStatus;
+            //order.OrderStatus = orderStatus;
 
             _context.Orders.Update(order);
             await _context.SaveChangesAsync();
